linux-imx: patch the type-c driver to boot in no-battery condition
These patches enable proper operation of the type-c port in drd and
drp modes, including the case where the system is not battery backed.

Critically, when the device is not self-powered, we can't let the
type-c driver tell the remote end to reset the bus, since that will
kill our power. This change adds a patch to do just this, and also
fixes up the device tree graph so drd will play nicely with the whole
setup.